requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
channels==4.0.0
daphne==4.0.0
fredapi==0.5.1
pandas==2.1.1
django-celery-beat==2.5.0
channels-redis==4.1.0
drf-yasg==1.21.7
flower==2.0.1
python-dotenv==1.0.0
gunicorn==21.2.0 